requirements.txt
langchain
langchain_community
langchain_core
chromadb
python-dotenv
pypdf
groq
langchain_groq
streamlit